requirements.txt
fastapi[all] #framework
uvicorn #servidor
sqlalchemy[asyncio]>=2.0.0 #ORM asincronico
asyncpg #optimizacion asincronica para postgreSQL
pydantic-settings #validacion .env
geoalchemy2 #PostGIS
alembic #control de versiones BD
python-jose[cryptography] #JWT
werkzeug #bcrypt
python-multipart #OAuth2
psycopg2-binary #sincrono
shapely # convertir coordenadas a objetos
openai # integracion de IA para el diagnostico